Adding ajax lookup on minimap for results (Fix #81)
Lookup is made on first moseover on marker.
If there is only one search result the label will be replaced with the
resource title and link directly to the resource.
If more than one; add number of results to the label and use the current
linking to the search results